Updated the code
I added the ability for the code to save the status of the drops to a separate file that can be modified. The file is status.txt. The bot reads the file in and determines what drops to get/skip. After watching a streamer for 123 minutes, the bot will save the position of the streamer (ie, link position 4 will save as status number 4) as True in status, indicating that the user has the drop and should be skipped in the future.